# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JagoEvent Ticketing API",
        "description": "REST API untuk aplikasi JagoEvent — marketplace tiket event. Digunakan oleh aplikasi mobile Flutter untuk browse events, order tickets, manage vendor, dan validasi tiket.",
        "contact": {
            "name": "JagoEvent Support",
            "email": "support@JagoEvent.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "JagoEvent API"
        }
    ],
    "paths": {
...
        "/vendors/user/{id}": {
            "get": {
                "tags": [
                    "Vendors"
                ],
                "summary": "Get vendor by user ID",
                "operationId": "a71c8301851f679c0f6e13297b79e86d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar vendor milik user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get vendor by user"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "My Event Organizer"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Jl. Sudirman No. 1"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "08123456789"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Jakarta"
                                                    },
                                                    "verify_status": {
                                                        "type": "string",
                                                        "example": "approved",
                                                        "enum": [
                                                            "pending",
                                                            "approved",
                                                            "rejected"
                                                        ]
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/vendor": {
            "post": {
                "tags": [
                    "Vendors"
                ],
                "summary": "Register sebagai vendor",
                "description": "Membuat vendor baru dan mengubah is_vendor user menjadi 1.",
                "operationId": "c44129d8bd2c65fb734eaaa6b6846a37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "name",
                                    "description",
                                    "location",
                                    "phone",
                                    "city"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "My Event Organizer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Event organizer terbaik"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Jl. Sudirman No. 1"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vendor berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vendor created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "location": {
                                                    "type": "string"
                                                },
                                                "phone": {
                                                    "type": "string"
                                                },
                                                "city": {
                                                    "type": "string"
                                                },
                                                "verify_status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Sanctum token (without Bearer prefix)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}